Update WebFlux samples to use Spring Boot
Fixes: gh-5411
