+++ samples/boot/hellowebflux-method/src/main/java/sample/SecurityConfig.java
