+++ samples/boot/hellowebfluxfn/src/main/java/sample/HelloWebfluxFnApplication.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2002-2017 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package sample;
+
+import static org.springframework.web.reactive.function.server.RequestPredicates.GET;
+import static org.springframework.web.reactive.function.server.RouterFunctions.route;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.reactive.function.server.RouterFunction;
+import org.springframework.web.reactive.function.server.ServerResponse;
+
+/**
+ * @author Rob Winch
+ * @since 5.0
+ */
+@SpringBootApplication
+public class HelloWebfluxFnApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(HelloWebfluxFnApplication.class, args);
+	}
+
+	@Bean
+	public RouterFunction<ServerResponse> routes(HelloUserController userController) {
+		return route(
+			GET("/"), userController::hello);
+	}
+}
