+++ samples/boot/webflux-form/src/main/resources/templates/index.html
