--- samples/javaconfig/hellowebflux-method/spring-security-samples-javaconfig-hellowebflux-method.gradle
@@ -1,16 +1,0 @@
-apply plugin: 'io.spring.convention.spring-sample'
-
-dependencies {
-	compile project(':spring-security-core')
-	compile project(':spring-security-config')
-	compile project(':spring-security-web')
-	compile 'com.fasterxml.jackson.core:jackson-databind'
-	compile 'io.projectreactor.ipc:reactor-netty'
-	compile 'org.springframework:spring-context'
-	compile 'org.springframework:spring-webflux'
-
-	testCompile project(':spring-security-test')
-	testCompile 'io.projectreactor:reactor-test'
-	testCompile 'org.skyscreamer:jsonassert'
-	testCompile 'org.springframework:spring-test'
-}
